move scanner.tasks.scan to scanner.scan
