+++ httpobs/scanner/local.py
@@ -1,4 +1,4 @@
-from httpobs.scanner.tasks import scan
+from httpobs.scanner import scan
 
 # for backwards compatibility, so consumers can continue to use httpobs.scanner.local.scan
 __all__ = ["scan"]
